Moved Plugin class into its own file
